requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-anymail==7.0.0
django-classy-tags==1.0.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-impersonate==1.5
django-mail-templated==2.6.5
django-rosetta==0.9.3
django-sekizai==1.1.0
django-stronghold==0.4.0
Faker==4.0.2
idna==2.9
model-bakery==1.1.0
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.3
Pillow==7.1.1
pluggy==0.13.1
polib==1.1.0
py==1.8.1
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==5.4.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
q==2.6
reportlab==3.5.42
requests==2.23.0
requests-oauthlib==1.3.0
sentry-sdk==0.14.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.8
wcwidth==0.1.9